agents/requirements.txt
anthropic>=0.25.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
fastapi>=0.111.0
uvicorn>=0.30.0
python-dotenv>=1.0.0
sentence-transformers>=3.0.0
numpy>=1.26.0
scikit-learn>=1.4.0
sqlalchemy>=2.0.0
apscheduler>=3.10.0
lxml>=5.2.0
feedparser>=6.0.0
python-slugify>=8.0.0
supabase>=2.0.0
